refactor: DatasetEditModal component
