+++ src/custom/ecospheres/components/BouquetDatasetList.vue
@@ -3,14 +3,15 @@ import { computed, ref, defineModel, type Ref } from 'vue'
 import { VueDraggableNext as draggable } from 'vue-draggable-next'
 
 import config from '@/config'
-import type { DatasetModalData } from '@/model/dataset'
-import { type DatasetProperties, Availability } from '@/model/topic'
+import { type DatasetProperties } from '@/model/topic'
 import { fromMarkdown } from '@/utils'
 import { isAvailable } from '@/utils/topic'
 
 import BouquetDatasetAccordionTitle from './BouquetDatasetAccordionTitle.vue'
 import BouquetDatasetCard from './BouquetDatasetCard.vue'
-import DatasetPropertiesFields from './forms/dataset/DatasetPropertiesFields.vue'
+import DatasetEditModal, {
+  type DatasetEditModalType
+} from './forms/dataset/DatasetEditModal.vue'
 
 const datasets = defineModel({
   type: Array<DatasetProperties>,
@@ -26,39 +27,14 @@ defineProps({
 
 const emits = defineEmits(['updateDatasets'])
 
+const modal: Ref<DatasetEditModalType | null> = ref(null)
 const isReorder = ref(false)
 const expandStore: Ref<{ [key: string]: string | null }> = ref({})
-const isModalOpen = ref(false)
 // make a copy for local reordering before save
 const originalDatasets = ref([...datasets.value])
-const modalData: Ref<DatasetModalData> = ref({
-  isValid: false,
-  mode: 'edit'
-})
 
 const expandedIds = computed(() => {
   return Object.keys(expandStore.value).filter((k) => !!expandStore.value[k])
-})
-
-const modalActions = computed(() => {
-  return [
-    {
-      label: 'Enregistrer',
-      disabled: !modalData.value.isValid,
-      onClick: ($event: PointerEvent) => {
-        $event.preventDefault()
-        submitModal(modalData.value)
-        closeModal()
-      }
-    },
-    {
-      label: 'Annuler',
-      secondary: true,
-      onClick: () => {
-        closeModal()
-      }
-    }
-  ]
 })
 
 const expandAll = () => {
@@ -75,44 +51,6 @@ const getAccordeonId = (index: number): string => {
   return `accordion_${index}`
 }
 
-const editDataset = (dataset: DatasetProperties, index: number) => {
-  // clone the object to enable cancellation
-  modalData.value = {
-    index,
-    dataset: { ...dataset },
-    isValid: false,
-    mode: 'edit'
-  }
-  isModalOpen.value = true
-}
-
-const addDataset = () => {
-  modalData.value = {
-    index: undefined,
-    dataset: {
-      title: '',
-      purpose: '',
-      availability: Availability.LOCAL_AVAILABLE,
-      uri: null,
-      id: null
-    },
-    isValid: false,
-    mode: 'create'
-  }
-  isModalOpen.value = true
-}
-
-const submitModal = (modalData: DatasetModalData) => {
-  if (modalData.dataset !== undefined) {
-    if (modalData.mode === 'create') {
-      datasets.value.push(modalData.dataset)
-    } else if (modalData.mode === 'edit' && modalData.index !== undefined) {
-      datasets.value[modalData.index] = modalData.dataset
-    }
-  }
-  emits('updateDatasets')
-}
-
 const removeDataset = (index: number) => {
   if (
     window.confirm(
@@ -125,10 +63,6 @@ const removeDataset = (index: number) => {
   }
 }
 
-const closeModal = () => {
-  isModalOpen.value = false
-}
-
 const saveOrder = () => {
   isReorder.value = false
   emits('updateDatasets')
@@ -137,6 +71,14 @@ const saveOrder = () => {
 const cancelReorder = () => {
   datasets.value = [...originalDatasets.value]
   isReorder.value = false
+}
+
+const addDataset = () => {
+  modal.value?.addDataset()
+}
+
+const editDataset = (dataset: DatasetProperties, index: number) => {
+  modal.value?.editDataset(dataset, index)
 }
 
 const triggerReorder = () => {
@@ -284,34 +226,13 @@ const triggerReorder = () => {
   </div>
 
   <!-- add/edit modal -->
-  <DsfrModal
-    v-if="isEdit && isModalOpen && modalData.dataset"
-    size="lg"
-    class="bouquet-dataset-modal"
-    :title="
-      modalData.mode === 'edit'
-        ? 'Éditer le jeu de données'
-        : 'Ajouter un jeu de données'
-    "
-    :opened="isModalOpen"
-    :actions="modalActions"
-    @close="closeModal"
-  >
-    <DatasetPropertiesFields
-      v-model:dataset-properties="modalData.dataset"
-      :already-selected-datasets="datasets"
-      @update-validation="(isValid: boolean) => modalData.isValid = isValid"
-    />
-  </DsfrModal>
+  <DatasetEditModal
+    v-if="isEdit"
+    ref="modal"
+    v-model="datasets"
+    @submit-modal="emits('updateDatasets')"
+  />
 </template>
-
-<style lang="scss">
-.bouquet-dataset-modal {
-  h1 {
-    margin-bottom: 1rem;
-  }
-}
-</style>
 
 <style scoped lang="scss">
 .ghost {
